% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bpr_control_data}
\alias{bpr_control_data}
\title{Synthetic control data for mpgex package}
\format{A list with 600 elements, where each element element is an L x 3
 matrix of observations, where:
\describe{
  \item{1st column}{locations of obseravtions}
  \item{2nd column}{total trials at corresponding locations}
  \item{3rd column}{number of successes at corresponding locations}
}}
\usage{
bpr_control_data
}
\description{
A synthetic dataset containinig 600 entries for control data.
}
\keyword{datasets}

