% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr.R
\name{bpr_derivative}
\alias{bpr_derivative}
\title{Derivative of the BPR log likelihood function}
\usage{
bpr_derivative(w, H, data)
}
\arguments{
\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{H}{The \code{L x M} matrix design matrix, where L is the number
of observations and M the number of basis functions.}

\item{data}{An \code{L x 2} matrix containing in the 1st column the number of
total reads and and in the 2nd the number of methylated reads. Each row
corresponds to each row of the design matrix.}
}
\value{
the derivative vector of the log likelihood w.r.t the vector of
coefficients w
}
\description{
\code{bpr_derivative} computes the derivative w.r.t the coefficients w of
 the Binomial distributed Probit regression log likelihood function.
}
\section{Mathematical formula}{

 The derivative of the Binomial distributed Probit regression log likelihood
 function w.r.t to w is computed by the following formula:
\deqn{log p(y | f, w) = \sum_{l=1}^{L} log Binom(m_{l} | t_{l}, \Phi(w^{t}h(x_{l})))}
}
\examples{
obj <- polynomial.object(M=2)
obs <- c(0,.2,.5)
des_mat <- design_matrix(obj, obs)
H <- des_mat$H
w <- c(.1,.1,.1)
data <- matrix(c(10,12,15,7,9,8), ncol=2)
der <- bpr_derivative(w, H, data)

}
\seealso{
\code{\link{bpr_likelihood}}, \code{\link{design_matrix}}
}

