% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_fdmm.R
\name{bpr_fdmm}
\alias{bpr_fdmm}
\title{Gibbs sampling algorithm for BPR mixture model}
\usage{
bpr_fdmm(x, K = 2, pi_k = NULL, w = NULL, basis = NULL,
  w_0_mean = NULL, w_0_cov = NULL, dir_a = NULL, gibbs_nsim = 5000,
  gibbs_burn_in = 1000, is_parallel = TRUE, no_cores = NULL,
  is_verbose = FALSE)
}
\arguments{
\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{K}{Integer denoting the number of clusters K.}

\item{pi_k}{Vector of length K, denoting the mixing proportions.}

\item{w}{A MxK matrix, where each column contains the basis function
coefficients for the corresponding cluster.}

\item{basis}{A 'basis' object. E.g. see \code{\link{polynomial.object}}}

\item{w_0_mean}{The prior mean hyperparameter for w}

\item{w_0_cov}{The prior covariance hyperparameter for w}

\item{dir_a}{The Dirichlet concentration parameter, prior over pi_k}

\item{gibbs_nsim}{Argument giving the number of simulations of the
Gibbs sampler.}

\item{gibbs_burn_in}{Argument giving the burn in period of the
Gibbs sampler.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_verbose}{Logical, print results during EM iterations}
}
\description{
\code{bpr_fdmm} implements the Gibbs sampling algorithm for performing
 clustering on DNA methylation profiles, where the observation model is the
 Binomial distributed Probit Regression function,
 \code{\link{bpr_likelihood}}.
}

