% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_gibbs.R
\name{bpr_gibbs.list}
\alias{bpr_gibbs.list}
\title{Gibbs sampling for the BPR model using list x}
\usage{
\method{bpr_gibbs}{list}(x, w_mle = NULL, basis = NULL,
  fit_feature = NULL, cpg_dens_feat = FALSE, w_0_mean = NULL,
  w_0_cov = NULL, gibbs_nsim = 100, gibbs_burn_in = 10,
  is_parallel = TRUE, no_cores = NULL, ...)
}
\arguments{
\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{w_mle}{A matrix of MLE estimates for the regression coefficients for
each genomic region of interest.}

\item{basis}{A 'basis' object. See \code{\link{polynomial.object}}}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{cpg_dens_feat}{Additional feature for the CpG density across the
promoter region.}

\item{w_0_mean}{The prior mean hyperparameter for w}

\item{w_0_cov}{The prior covariance hyperparameter for w}

\item{gibbs_nsim}{Optional argument giving the number of simulations of the
Gibbs sampler.}

\item{gibbs_burn_in}{Optional argument giving the burn in period of the
Gibbs sampler.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 2.}

\item{...}{Additional parameters}
}
\value{
A list containing the following elements:
\itemize{
 \item{ \code{W_opt}: An Nx(M+1) matrix with the optimized parameter values.
   Each row of the matrix corresponds to each element of the list x. The
   columns are of the same length as the parameter vector w (i.e. number
   of basis functions).
 }
 \item{ \code{Mus}: An N x M matrix with the RBF centers if basis object is
  \code{\link{rbf.object}}, otherwise NULL.}
 \item{ \code{basis}: The basis object.
 }
 \item{ \code{w}: The initial values of the parameters w.
 }
 \item{ \code{x_extrema}: The min and max values of each promoter region.
 }
}
}
\description{
\code{bpr_gibbs.list} computes the posterior of the BPR model using auxiliary
variable approach. Since we cannot compute the posterior analytically, a
Gibbs sampling scheme is used. This method calls
\code{\link{bpr_gibbs.matrix}} to process each element of the list.
}
\examples{
ex_data <- bpr_data
basis <- rbf.object(M=3)
out_opt <- bpr_gibbs(x = ex_data, is_parallel = FALSE, basis = basis)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_optim.matrix}}
}

