% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_gibbs.R
\name{bpr_gibbs.matrix}
\alias{bpr_gibbs.matrix}
\title{Gibbs sampling for the BPR model using list x}
\usage{
\method{bpr_gibbs}{matrix}(x, w_mle = NULL, basis = NULL,
  fit_feature = NULL, cpg_dens_feat = FALSE, w_0_mean = NULL,
  w_0_cov = NULL, gibbs_nsim = 20, gibbs_burn_in = 10, ...)
}
\arguments{
\item{x}{An L x 3 matrix of observations, where 1st column contains the
locations. The 2nd and 3rd columns contain the total trials and number of
successes at the corresponding locations, repsectively.}

\item{w_mle}{A vector of parameters (i.e. coefficients of the basis functions)
containing the MLE estimates.}

\item{basis}{A 'basis' object. See \code{\link{polynomial.object}}}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{cpg_dens_feat}{Additional feature for the CpG density across the
promoter region.}

\item{w_0_mean}{The prior mean hyperparameter for w}

\item{w_0_cov}{The prior covariance hyperparameter for w}

\item{gibbs_nsim}{Optional argument giving the number of simulations of the
Gibbs sampler.}

\item{gibbs_burn_in}{Optional argument giving the burn in period of the
Gibbs sampler.}

\item{...}{Additional parameters}
}
\value{
A list containing the following elements:
\itemize{
 \item{ \code{w_opt}: Optimized values for the coefficient vector w.
   The length of the result is the same as the length of the vector w.
 }
 \item{ \code{basis}: The basis object.
 }
}
}
\description{
\code{bpr_gibbs.matrix} computes the posterior of the BPR model using auxiliary
variable approach. Since we cannot compute the posterior analytically, a
Gibbs sampling scheme is used.
}
\examples{
basis <- polynomial.object(M=2)
w <- c(0.1, 0.1, 0.1)
w_0_mean <- rep(0, length(w))
w_0_cov <- diag(10, length(w))
data <- bpr_data[[1]]
out_opt <- bpr_gibbs(x = data, w_mle = w, w_0_mean = w_0_mean,
                     w_0_cov = w_0_cov, basis = basis)

basis <- polynomial.object(M=0)
w <- c(0.1)
w_0_mean <- rep(0, length(w))
w_0_cov <- diag(10, length(w))
data <- bpr_data[[1]]
out_opt <- bpr_gibbs(x = data, w_mle = w, w_0_mean = w_0_mean,
                     w_0_cov = w_0_cov, basis = basis)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_optim.list}}
}

