% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_gibbs.R
\name{bpr_gibbs}
\alias{bpr_gibbs}
\title{Generic function for performing Gibbs sampling on BPR model}
\usage{
bpr_gibbs(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Additional parameters}
}
\description{
\code{bpr_optim} is a generic function which calles the appropriate methods
depending on the class of the object \code{x}. Object \code{x} can be either
a \code{\link{list}} or a \code{\link{matrix}}.
}
\examples{
data <- bpr_data
out_opt <- bpr_gibbs(x = data, is_parallel = FALSE)

}
\seealso{
\code{\link{bpr_gibbs.list}}, \code{\link{bpr_gibbs.matrix}}
}

