% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimx.R
\name{bpr_optimx.list}
\alias{bpr_optimx.list}
\title{Optimization method for the BPR log likelihood function using a list X}
\usage{
\method{bpr_optimx}{list}(X, w = NULL, basis = NULL, method = "CG",
  itnmax = 200)
}
\description{
Optimization method for the BPR log likelihood function using a list X
}

