% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimx.R
\name{bpr_optimx.matrix}
\alias{bpr_optimx.matrix}
\title{Optimization method for the BPR log likelihood function using a matrix x}
\usage{
\method{bpr_optimx}{matrix}(x, w = NULL, basis = NULL, method = "CG",
  itnmax = 200, ...)
}
\arguments{
\item{x}{An L x 3 matrix of observations, where 1st column contains the
locations. The 2nd and 3rd columns contain the total trials and number of
successes at the corresponding locations, repsectively.}

\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{basis}{A 'basis' object. See \code{\link{polynomial.object}}}

\item{method}{The optimization method to be used. See \code{\link[optimx]{optimx}}
for possible methods. Default is 'CG'.}

\item{itnmax}{Optional argument giving the maximum number of iterations for
the corresponding method. See \code{\link[optimx]{optimx}} for details.}

\item{...}{Additional parameters}
}
\value{
Optimized values for the parameter/coefficient vector w. The length
 of the result is the same as the length of the vector w.
}
\description{
\code{bpr_optimx.matrix} computes the maximum log likelihood of the BPR
function. Since it cannot be evaluated analytically, an optimization
procedure is used. The \code{\link[optimx]{optimx}} packages is used for
performing optimization.
}
\seealso{
\code{\link{bpr_optimx}}, \code{\link{bpr_optimx.list}}
}

