% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimx.R
\name{bpr_optimx}
\alias{bpr_optimx}
\title{Generic function for optimizing the BPR log likelihood function}
\usage{
bpr_optimx(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Additional parameters}
}
\description{
This is a generic function which calles the appropriate methods depending
on the class of the object \code{x}. Currently only the 'polynomial' class
is implemented.
}
\examples{
obj <- polynomial.object(M=2)
obs <- c(0,.2,.5)
H <- design_matrix(obj, obs)

}
\seealso{
\code{\link{bpr_optimx.list}}, \code{\link{bpr_optimx.matrix}}
}

