% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimize.R
\name{bpr_scaled_optim}
\alias{bpr_scaled_optim}
\title{Optimization method for the scaled BPR NLL function using list x}
\usage{
bpr_scaled_optim(x, W_contr, w = NULL, basis = NULL, method = "CG",
  itnmax = 100, ...)
}
\arguments{
\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{W_contr}{An N x M matrix containing the basis coefficients of a
previously fitted profile. Each row will be used as a constant vector
which will be scaled when performing optimization on the new profiles.}

\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{basis}{A 'basis' object. See \code{\link{polynomial.object}}}

\item{method}{The optimization method to be used. See \code{\link[stats]{optim}}
for possible methods. Default is 'CG'.}

\item{itnmax}{Optional argument giving the maximum number of iterations for
the corresponding method. See \code{\link[stats]{optim}} for details.}

\item{...}{Additional parameters}
}
\value{
A list containing the following elements:
\itemize{
 \item{ \code{W_opt}: An Nx(M+1) matrix with the optimized parameter values.
   Each row of the matrix corresponds to each element of the list x. The
   columns are of the same length as the parameter vector w (i.e. number
   of basis functions).
 }
 \item{ \code{Mus}: An NxM matrix with the RBF centers if basis object is
  \code{\link{rbf.object}}, otherwise NULL.}
 \item{ \code{basis}: The basis object.
 }
 \item{ \code{w}: The initial values of the parameters w.
 }
}
}
\description{
\code{bpr_scaled_optim} minimizes the negative log likelihood of the scaled
BPR function. Since it cannot be evaluated analytically, an optimization
procedure is used. The \code{\link[stats]{optim}} packages is used for
performing optimization. This method calls \code{\link{bpr_optim.matrix}}
to process each element of the list.
}
\seealso{
\code{\link{bpr_optim.matrix}}
}

