% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_mpgex.R
\name{differential_mpgex}
\alias{differential_mpgex}
\title{Predict differntial gene expression from differential methylation}
\usage{
differential_mpgex(formula = NULL, X, Y, train_ind = NULL, basis = NULL,
  w = NULL, train_perc = 0.7, opt_method = "CG", opt_itnmax = 100,
  lambda = 0)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression method is used.}

\item{X}{The binomial distributed observations. A list containing two lists
for control and treatment samples. Each of the two lists has a nested list
where each element is an L x 3 dimensional matrix.}

\item{Y}{Corresponding gene expression data for each element of the list X}

\item{train_ind}{Optional argument containing the indices for splitting the
dataset to train and test sets.}

\item{basis}{Optional basis function object, default is
\code{\link{polynomial.object}}}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{opt_method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{opt_itnmax}{Optional parameter for defining the max number of
iterations of the optimization procedure, see \code{\link[stats]{optim}}.}

\item{lambda}{Regularization term when performing Basis Linear Model fitting}
}
\value{
An mpgex object consisting of the following elements: ....
}
\description{
\code{differential_mpgex} is a function that wraps all the necessary
subroutines for performing predictions on differential gene expression.
Initially, it optimizes the parameters of the basis functions so as to
learn the methylation profiles for the control case and the treatment case.
Then the two learned profiles for each promoter region are subtracted and
new coefficients showing the difference between the two profiles are learnt
using the Basis Linear Model. These coefficients are given as input features
for performing linear regression in order to predict/regress the
corresponding differential gene expression data.
}
\examples{
obs <- list(control = bpr_control_data, treatment = bpr_treatment_data)
Y   <- list(control = gex_control_data, treatment = gex_treatment_data)
basis <- rbf.object(M = 5, gamma = 0.3)
out   <- differential_mpgex(X = obs, Y = Y, basis = basis, lambda = 1e-02)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_likelihood}},
 \code{\link{blm}}, \code{\link{design_matrix}}
}

