% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{eval_polyn_func}
\alias{eval_polyn_func}
\title{Evaluate polynomial function}
\usage{
eval_polyn_func(x, w)
}
\arguments{
\item{x}{Input / observation data.}

\item{w}{Vector of length M, containing the coefficients of an Mth-order
basis function.}
}
\value{
The polynomial function values.
}
\description{
Method for evaluating the the polynomial function of degree M for
observation data x and coefficients w.
}
\examples{
x <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
out <- eval_polyn_func(x, w)

}

