% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_prob_rbf_func}
\alias{eval_prob_rbf_func}
\title{Radia basis function of degree M.
X is the observation}
\usage{
eval_prob_rbf_func(x, w, basis, mus)
}
\description{
Radia basis function of degree M.
X is the observation
}

