% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{eval_rbf_func}
\alias{eval_rbf_func}
\title{Evaluate rbf function}
\usage{
eval_rbf_func(x, w, basis, mus)
}
\arguments{
\item{x}{Input / observation data.}

\item{w}{Vector of length M, containing the coefficients of an Mth-order
basis function.}

\item{basis}{The basis object.}

\item{mus}{The centers of the rbf function.}
}
\value{
The rbf function values.
}
\description{
Method for evaluating the rbf function with M basis for observation
data x and coefficients w.
}
\examples{
x <- c(1,2,3)
w <- c(0.1, 0.3, -0.6)
basis <- rbf.object(M = 2)
mus <- c(2,2.5)
out <- eval_rbf_func(x, w, basis, mus)

}

