% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diff_meth.R
\name{fit_diff_meth}
\alias{fit_diff_meth}
\title{Fit differential methylation profiles using BLM}
\usage{
fit_diff_meth(control, treatment, diff_basis, lambda, evals = 50)
}
\arguments{
\item{control}{The optimized values for the control samples.}

\item{treatment}{The optimized values for the treatment samples.}

\item{diff_basis}{The basis object for differential methylation.}

\item{lambda}{The regularization parameter when fitting the BLM.}

\item{evals}{Integer denoting the number of evaluation points.}
}
\value{
The differential methylation profiles and the corresponding fitted
 coefficients from the BLM.
}
\description{
Function for fitting the differential methylation profiles and returning
the fitted coefficients using the Basis Linear Model.
}

