% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gex_control_data}
\alias{gex_control_data}
\title{Synthetic control data for mpgex package}
\format{A vector of length 600}
\usage{
gex_control_data
}
\description{
Corresponding gene expression data for the 'bpr_control_data'
}
\keyword{datasets}

