% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gex_treatment_data}
\alias{gex_treatment_data}
\title{Synthetic treatment data for mpgex package}
\format{A vector of length 600}
\usage{
gex_treatment_data
}
\description{
Corresponding gene expression data for the 'bpr_treatment_data'
}
\keyword{datasets}

