% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_diff_meth.R
\name{learn_diff_meth}
\alias{learn_diff_meth}
\title{Learn differential methylation profiles using BLM}
\usage{
learn_diff_meth(control, treatment, diff_basis, fit_feature = NULL,
  lambda = 0, x_evals = 50)
}
\arguments{
\item{control}{The optimized values for the control samples.}

\item{treatment}{The optimized values for the treatment samples.}

\item{diff_basis}{The basis object for differential methylation.}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{lambda}{The regularization parameter when fitting the BLM.}

\item{x_evals}{Integer denoting the number of evaluation points.}
}
\value{
The differential methylation profiles and the corresponding fitted
 coefficients from the BLM model.
}
\description{
\code{learn_diff_meth} learns differential methylation profiles and returns
 the fitted coefficients using the Basis Linear Model.
}

