% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex.R
\docType{package}
\name{mpgex}
\alias{mpgex}
\alias{mpgex-package}
\title{\code{mpgex}: A package for predicting gene expression from methylation profiles}
\description{
\code{mpgex}: A package for predicting gene expression from methylation profiles
}

