% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_cluster_bayes.R
\name{mpgex_cluster_bayes}
\alias{mpgex_cluster_bayes}
\title{Cluster similar methylation profiles using Gibbs smapling}
\usage{
mpgex_cluster_bayes(x, K = 3, pi_k = NULL, w = NULL, basis = NULL,
  w_0_mean = NULL, w_0_cov = NULL, gibbs_nsim = 5000,
  gibbs_burn_in = 1000, opt_method = "CG", opt_itnmax = 50,
  is_parallel = TRUE, no_cores = NULL, is_verbose = FALSE)
}
\arguments{
\item{x}{A list of elements of length N, where each element is an L x 3
matrix of observations, where 1st column contains the locations. The 2nd
and 3rd columns contain the total trials and number of successes at the
corresponding locations, repsectively.}

\item{K}{Integer denoting the number of clusters K.}

\item{pi_k}{Vector of length K, denoting the mixing proportions.}

\item{w}{A MxK matrix, where each column contains the basis function
coefficients for the corresponding cluster.}

\item{basis}{A 'basis' object. E.g. see \code{\link{polynomial.object}}}

\item{w_0_mean}{The prior mean hyperparameter for w}

\item{w_0_cov}{The prior covariance hyperparameter for w}

\item{gibbs_nsim}{Argument giving the number of simulations of the
Gibbs sampler.}

\item{gibbs_burn_in}{Argument giving the burn in period of the
Gibbs sampler.}

\item{opt_method}{Optimization method used.}

\item{opt_itnmax}{Optimization iterations for obtaining the initial EM
parameter values.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_verbose}{Logical, print results during EM iterations}
}
\description{
\code{mpgex_cluster} is a wrapper function that clusters similar methylation
profiles using the Gibbs sampling algorithm. Initially, it performs parameter
checking, and initializes main parameters, such as mixing proportions, basis
function coefficients, then Gibbs is applied and finally model selection
metrics are calculated, such as BIC and AIC.
}
\examples{
set.seed(123)
ex_data <- bpr_data
mpgex_clust <- mpgex_cluster_bayes(x = ex_data, gibbs_nsim = 20, gibbs_burn_in = 10,
                                   is_parallel = FALSE, opt_itnmax = 10)

}

