% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_differential_regr.R
\name{mpgex_differential_regr}
\alias{mpgex_differential_regr}
\title{Predict simple differntial gene expression from differential methylation}
\usage{
mpgex_differential_regr(formula = NULL, x, y, model_name = "svm",
  w = NULL, basis = NULL, train_ind = NULL, train_perc = 0.7,
  fit_feature = NULL, opt_method = "CG", opt_itnmax = 100,
  is_parallel = TRUE, no_cores = NULL, is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression method is used.}

\item{x}{The binomial distributed observations. A list containing two lists
for control and treatment samples. Each of the two lists has a nested list
where each element is an L x 3 dimensional matrix.}

\item{y}{The gene expression data. A list containing two vectors for control
and treatment samples.}

\item{model_name}{A charcter denoting the regression model.}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{basis}{Optional basis function object, default is
\code{\link{polynomial.object}}}

\item{train_ind}{Optional vector containing the indices for the
train set.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{opt_method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{opt_itnmax}{Optional parameter for defining the max number of
iterations of the optimization procedure, see \code{\link[stats]{optim}}.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
An mpgex object consisting of the following elements:
}
\description{
\code{mpgex_differential_regr} is a function that wraps all the necessary
subroutines for performing predictions on differential gene expression.
Initially, it optimizes the parameters of the basis functions so as to
learn the methylation profiles for the control case and the treatment case.
Then the two learned profiles for each promoter region are subtracted and
new coefficients showing the difference between the two profiles are learned
using the Basis Linear Model. These coefficients are given as input features
for performing linear regression in order to predict/regress the
corresponding differential gene expression data.
}
\examples{
obs <- list(control = bpr_control_data, treatment = bpr_treatment_data)
y   <- list(control = gex_control_data, treatment = gex_treatment_data)
basis <- rbf.object(M = 3)
set.seed(1234)
out   <- mpgex_differential_regr(x = obs, y = y, basis = basis,
                                 is_parallel = FALSE, opt_itnmax = 5)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_likelihood}},
 \code{\link{blm}}, \code{\link{polynomial.object}},
 \code{\link{rbf.object}}
}

