% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_regr_bayes.R
\name{mpgex_regr_bayes}
\alias{mpgex_regr_bayes}
\title{Predict gene expression from methylation profiles using Gibbs Sampling}
\usage{
mpgex_regr_bayes(formula = NULL, x, y, model_name = "svm", w = NULL,
  basis = NULL, train_ind = NULL, train_perc = 0.7, fit_feature = NULL,
  cpg_dens_feat = FALSE, w_0_mean = NULL, w_0_cov = NULL,
  gibbs_nsim = 5000, gibbs_burn_in = 1000, opt_method = "CG",
  opt_itnmax = 50, is_parallel = TRUE, no_cores = NULL,
  is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the SVM regression method is used.}

\item{x}{The binomial distributed observations, which has to be a list
where each element is an L x 3 dimensional matrix.}

\item{y}{Corresponding gene expression data for each element of the list x}

\item{model_name}{A charcter denoting the regression model.}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{basis}{Optional basis function object, default is
\code{\link{polynomial.object}}}

\item{train_ind}{Optional vector containing the indices for the
train set.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{fit_feature}{Additional feature on how well the profile fits the
methylation data.}

\item{cpg_dens_feat}{Additional feature for the CpG density across the
promoter region.}

\item{w_0_mean}{The prior mean hyperparameter for w}

\item{w_0_cov}{The prior covariance hyperparameter for w}

\item{gibbs_nsim}{Optional argument giving the number of simulations of the
Gibbs sampler.}

\item{gibbs_burn_in}{Optional argument giving the burn in period of the
Gibbs sampler.}

\item{opt_method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{opt_itnmax}{Optional parameter for defining the max number of
iterations of the optimization procedure, see \code{\link[stats]{optim}}.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 1.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
An mpgex object consisting of the following elements:
}
\description{
\code{mpgex_regr} is a function that wraps all the necessary subroutines
for performing predictions on gene expressions. Initially, it computes the
posterior distribution, using Gibbs sampling, of the parameters of the basis
functions so as to learn the methylation profiles. Then uses the learned
parameters / coefficients of the basis functions as input features for
performing linear regression in order to predict/regress the corresponding
gene expression data.
}
\examples{
obs <- bpr_data
y   <- gex_data
basis <- rbf.object(M = 5)
out   <- mpgex_regr_bayes(x = obs, y = y, basis = basis, is_parallel = FALSE,
                    gibbs_nsim = 20, gibbs_burn_in = 10)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_likelihood}},
 \code{\link{polynomial.object}}, \code{\link{rbf.object}},
 \code{\link{design_matrix}}
}

