% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition_data}
\alias{partition_data}
\title{Partition data in train and test set}
\usage{
partition_data(X, Y, train_ind = NULL, train_perc = 0.7)
}
\arguments{
\item{X}{Input / Independent variables}

\item{Y}{Dependent variables}

\item{train_ind}{Index of traininig data, if NULL a random one is generated.}

\item{train_perc}{Percentage of training data when partitioning.}
}
\value{
A list containing the train, test data and index of training data.
}
\description{
\code{partition_data} partition data randomly in train and test sets.
}
\examples{
X <- matrix(c(-20, 0, 15, 20, 20, 10, 4, 5), ncol=2)
Y <- c(2, 5, 0, 8)
data <- partition_data(X, Y)

##
X <- matrix(c(-20, 0, 15, 20, 20, 10, 4, 5), ncol=2)
Y <- c(2, 5, 0, 8)
train_ind <- c(1,4)
data <- partition_data(X, Y, train_ind)

}

