% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.diff_meth}
\alias{plot.diff_meth}
\title{Plot method for differential methylation}
\usage{
\method{plot}{diff_meth}(x, obs, w, ...)
}
\arguments{
\item{x}{The basis object.}

\item{obs}{An L x 3 matrix containing the observations data.}

\item{w}{Vector of length M, containing the fitted parameters / coefficients
for the M basis functions.}

\item{...}{Additional parameters}
}
\description{
Plot data using the differntial methylation data
}

