% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.rbf}
\alias{plot.rbf}
\title{S3 plot method for rbf objects}
\usage{
\method{plot}{rbf}(x, dataset, w, mus, ...)
}
\arguments{
\item{x}{Object type, which actually is the basis object.}

\item{dataset}{An L x 3 matrix containing the observations data.}

\item{w}{Vector of length M, containing the fitted parameters / coefficients
for the M basis functions.}

\item{mus}{The centers of the rbf function.}

\item{...}{Additional parameters}
}
\description{
Plot data using the rbf function as the response to observations x
}

