% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_diff_mpgex}
\alias{plot_diff_mpgex}
\title{Plot method for differential methylation}
\usage{
plot_diff_mpgex(x, t, ...)
}
\arguments{
\item{x}{Object of class 'diff_mpgex'}

\item{t}{Region to plot}
}
\description{
Plot data using the differntial methylation data
}

