% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fitted_profiles}
\alias{plot_fitted_profiles}
\title{Plot fit of methylation profiles across a region}
\usage{
plot_fitted_profiles(region, X, fit_prof, fit_mean = NULL,
  title = "Gene promoter", ...)
}
\arguments{
\item{region}{Promoter region number}

\item{X}{Methylation data observations}

\item{fit_prof}{Fitted profiles}

\item{fit_mean}{Fitted mean function}

\item{title}{Title of the plot}

\item{...}{Additional parameters}
}
\description{
Plot fit of methylation profiles across a region
}

