% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{polynomial.object}
\alias{polynomial.object}
\title{Creates a polynomial object}
\usage{
polynomial.object(M = 1)
}
\arguments{
\item{M}{The degree of the polynomial object that will be created.}
}
\value{
An object of class 'polynomial'
}
\description{
\code{polynomial.object} creates objects of class "polynomial". These
objects can be used as input to complex functions in order to perform
computation depending on the basis function. E.g see
\code{\link{design_matrix}} generic function.
}
\examples{
obj <- polynomial.object(M=2)
(obj)

}
\seealso{
\code{\link{design_matrix}}
}

