% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_gex.R
\name{predict_gex}
\alias{predict_gex}
\title{Predict gene expression data from methylation profiles}
\usage{
predict_gex(formula = NULL, model_name = "lm", train, test,
  is_summary = TRUE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression model is used.}

\item{model_name}{A charcter denoting the regression model}

\item{train}{The training data.}

\item{test}{The test data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements:
\itemize{
 \item \code{formula} the formula that was used.
 \item \code{gex_model} the fitted model.
 \item \code{train_pred} the training predicted values.
 \item \code{test_pred} the test predicted values.
 \item \code{train_errors} the training error metrics.
 \item \code{test_errors} the test error metrics.
}
}
\description{
\code{predict_gex} predicts gene expression data from methylation profiles,
 using linear regression. The same method can be used to predict
 differential expression from differential methylation profiles.
}
\seealso{
\code{\link{calculate_errors}}
}

