% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_mpgex.R
\name{wrapper_mpgex}
\alias{wrapper_mpgex}
\title{Predict gene expression from methylation profiles}
\usage{
wrapper_mpgex(formula = NULL, X, Y, model_name = "lm", train_ind = NULL,
  basis = NULL, w = NULL, train_perc = 0.7, method = "CG",
  itnmax = 100)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression method is used.}

\item{X}{The binomial distributed observations, which has to be a list
where each element is an L x 3 dimensional matrix.}

\item{Y}{Corresponding gene expression data for each element of the list X}

\item{model_name}{A charcter denoting the regression model}

\item{train_ind}{Optional argument containing the indices for splitting the
dataset to train and test sets.}

\item{basis}{Optional basis function object, default is
\code{\link{polynomial.object}}}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{itnmax}{Optional parameter for defining the max number of iterations
of the optimization procedure, see \code{\link[stats]{optim}}.}
}
\value{
An mpgex object consisting of the following elements: ...
}
\description{
\code{wrapper_mpgex} is a function that wraps all the necessary subroutines
for performing predictions on gene expressions. Initially, it optimizes the
parameters of the basis functions so as to learn the methylation profiles.
Then uses the learned parameters / coefficients of the basis functions as
input features for performing linear regression in order to predict/regress
the corresponding gene expression data.
}
\examples{
obs <- bpr_data
Y   <- gex_data
basis <- rbf.object(M = 5, gamma = 0.3)
out   <- wrapper_mpgex(X = obs, Y = Y, basis = basis)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_likelihood}},
 \code{\link{polynomial.object}}, \code{\link{rbf.object}},
 \code{\link{design_matrix}}
}

