% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_optimx.R
\name{bpr_optimx}
\alias{bpr_optimx}
\title{Optimization method for the BPR log likelihood function}
\usage{
bpr_optimx(X, w = NULL, basis = NULL, method = "CG", itnmax = 200)
}
\description{
Optimization method for the BPR log likelihood function
}

