% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_mpgex.R
\name{differential_mpgex}
\alias{differential_mpgex}
\title{Predict differntial gene expression from differential methylation}
\usage{
differential_mpgex(formula = NULL, X, Y, train_ind = NULL, basis = NULL,
  w = NULL, train_perc = 0.7, method = "CG", itnmax = 100)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} needed when
calling the \code{\link[stats]{lm}} function for performing linear
regression. If NULL, the simple linear regression method is used.}

\item{X}{The binomial distributed observations, which has to be a list
where each element is an L x 3 dimensional matrix.}

\item{Y}{Corresponding gene expression data for each element of the list X}

\item{train_ind}{Optional argument containing the indices for splitting the
dataset to train and test sets.}

\item{basis}{Optional basis function object, default is
\code{\link{polynomial.object}}}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{train_perc}{Optional parameter for defining the percentage of the
dataset to be used for training set, the remaining will be the test set.}

\item{method}{Parameter for defining the method to be used in the
optimization procedure, see \code{\link[stats]{optim}}.}

\item{itnmax}{Optional parameter for defining the max number of iterations
of the optimization procedure, see \code{\link[stats]{optim}}.}
}
\value{
An mpgex object consisting of the following elements:
}
\description{
\code{differential_mpgex} is a function that wraps all the necessary
subroutines for performing predictions on differential gene expression.
Initially, it optimizes the parameters of the basis functions so as to
learn the methylation profiles for the control case. Then uses these
parameters as constants and learns the methylation profiles of the treatment
as a scaled version of the control profiles. The coefficients showing
the difference between the two profiles is given as input features for
performing linear regression in order to predict/regress
the corresponding differential gene expression data.
}
\examples{
obs <- list(control = bpr_control_data, treatment = bpr_treatment_data)
Y   <- list(control = gex_control_data, treatment = gex_treatment_data)
basis <- rbf.object(M = 5, gamma = 0.3)
out   <- differential_mpgex(X = obs, Y = Y, basis = basis)

}
\seealso{
\code{\link{bpr_optim}}, \code{\link{bpr_likelihood}},
 \code{\link{polynomial.object}}, \code{\link{rbf.object}},
 \code{\link{design_matrix}}
}

