% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_prob_polyn_func}
\alias{eval_prob_polyn_func}
\title{Polynomial function of degree M.
X is the input vector and w are the coefficients of the
polynomial whose length is M + 1}
\usage{
eval_prob_polyn_func(x, w)
}
\description{
Polynomial function of degree M.
X is the input vector and w are the coefficients of the
polynomial whose length is M + 1
}

