% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgex_package.R
\docType{package}
\name{mpgex}
\alias{mpgex}
\alias{mpgex-package}
\title{\code{mpgex}: Package for predicting gene expression from methylation profiles}
\description{
\code{mpgex}: Package for predicting gene expression from methylation profiles
}
\section{\code{Base} functions}{

\code{\link{polynomial_basis}}
}

