% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition_data}
\alias{partition_data}
\title{Partition data in train and test set}
\usage{
partition_data(X, Y, train_perc = 0.7)
}
\arguments{
\item{X}{Input / Independent variables}

\item{Y}{Dependent variables}

\item{train_perc}{Percentage of training data when partitioning.}
}
\value{
A list containing the train and test data.
}
\description{
\code{partition_data} partition data randomly in train and test sets.
}
\examples{
X <- matrix(c(-20, 0, 15, 20), ncol=2)
Y <- c(2,5)
data <- partition_data(X, Y)

}

