% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.polynomial}
\alias{plot.polynomial}
\title{S3 plot method for polynomial objects}
\usage{
\method{plot}{polynomial}(x, dataset, w, ...)
}
\arguments{
\item{x}{Object type, which actually is the basis object.}

\item{dataset}{An L x 3 matrix containing the observations data.}

\item{w}{Vector of length M, containing the fitted parameters / coefficients
for the M basis functions.}

\item{...}{Additional parameters}
}
\description{
Plot data using the polynomial function as the response to observations x
}

